requirements.txt
fastapi>=0.115.0
uvicorn[standard]==0.27.0
sqlalchemy>=2.0.36
pydantic>=2.10.0
pydantic-settings==2.1.0
python-multipart==0.0.6
pyjwt==2.8.0
email-validator==2.1.1